COB: implemented StartBuilding
heading and pitch for this function are passed around as radians
in the engine and only in CCobInstance coverted to TA angles.

for this I needed to introduce GetHeadingFromVectorF and GetHAndPFromVectorF
(float versions of similarly named -but without F suffix- short int versions)

+++ rts/Sim/Units/COB/LuaUnitScript.cpp
@@ -332,7 +332,7 @@ void CLuaUnitScript::TransportDrop(CUnit* unit, const float3& pos)
 }
 
 
-void CLuaUnitScript::StartBuilding(int heading, int pitch)
+void CLuaUnitScript::StartBuilding(float heading, float pitch)
 {
 }
